Start refactoring cache
