+++ lib/cache/mongo-cache.js
@@ -2,10 +2,10 @@
 
 var Promise = require('bluebird');
 
-var metrics = require('./metrics/cache');
-var Response = require('./models/Response');
-var CacheItem = require('./models/CacheItem');
-var settings = require('./settings');
+var metrics = require('../metrics/cache');
+var Response = require('../models/Response');
+var CacheItem = require('../models/CacheItem');
+var settings = require('../settings');
 
 var CACHE_NAME = 'tile-cache-' + settings.name;
