+++ lib/cache/s3-cache-features.js
@@ -3,45 +3,24 @@
 var __ = require('lodash');
 var async = require('async');
 var concat = require('concat-stream');
+var knex = require('knex');
 var projector = require("nodetiles-core").projector;
 var Promise = require('bluebird');
 var xml2js = require('xml2js');
 
-var metrics = require('./metrics/cache');
-var Response = require('./models/Response');
-var settings = require('./settings');
-
-var knex = require('knex');
-
-// XXX TODO -- should this connection be created at the app level?
-var pg = knex({
-  client: 'pg',
-  connection: settings.psqlConnectionString
-});
-
+var metrics = require('../metrics/cache');
+var Response = require('../models/Response');
+var settings = require('../settings');
+var util = require('./cache-util');
 
 var S3_CUSTOM_CACHE_HEADER = 'x-amz-meta-count';
 
-function tile2long(x, z) {
-  return (x/Math.pow(2,z)*360-180);
-}
-
-function tile2lat(y, z) {
-  var n=Math.PI-2*Math.PI*y/Math.pow(2,z);
-  return (180 / Math.PI * Math.atan(0.5 * (Math.exp(n) - Math.exp(-n))));
-}
-
-function tileToBounds(tile) {
-  var sw = [tile2long(tile[1], tile[0]), tile2lat(tile[2] + 1, tile[0])];
-  var ne = [tile2long(tile[1] + 1, tile[0]), tile2lat(tile[2], tile[0])];
-  return [sw, ne];
-}
-
 module.exports = function setup(options) {
   var s3client = options.s3client;
+  var pg = options.pgClient;
 
   function makeEnvelope(tile) {
-    var bounds = tileToBounds(tile);
+    var bounds = util.tileToBounds(tile);
     var west = bounds[0][0];
     var south = bounds[0][1];
     var east = bounds[1][0];
