+++ lib/cache/s3-cache.js
@@ -4,30 +4,17 @@ var concat = require('concat-stream');
 var Promise = require('bluebird');
 var xml2js = require('xml2js');
 
-var metrics = require('./metrics/cache');
-var Response = require('./models/Response');
-var settings = require('./settings');
+var metrics = require('../metrics/cache');
+var Response = require('../models/Response');
+var settings = require('../settings');
+var util = require('./cache-util');
 
-function tile2long(x, z) {
-  return (x/Math.pow(2,z)*360-180);
-}
-
-function tile2lat(y, z) {
-  var n=Math.PI-2*Math.PI*y/Math.pow(2,z);
-  return (180 / Math.PI * Math.atan(0.5 * (Math.exp(n) - Math.exp(-n))));
-}
-
-function tileToBounds(tile) {
-  var sw = [tile2long(tile[1], tile[0]), tile2lat(tile[2] + 1, tile[0])];
-  var ne = [tile2long(tile[1] + 1, tile[0]), tile2lat(tile[2], tile[0])];
-  return [sw, ne];
-}
 
 module.exports = function setup(options) {
   var s3client = options.s3client;
 
   function setupQuery(survey, tile, deleted) {
-    var bounds = tileToBounds(tile);
+    var bounds = util.tileToBounds(tile);
     var west = bounds[0][0];
     var south = bounds[0][1];
     var east = bounds[1][0];
