+++ lib/controllers/tiles.js
@@ -11,7 +11,7 @@ var CacheItem = require('../models/CacheItem');
 
 var layer = require('../layer.js');
 var metrics = require('../metrics/pipeline');
-var responsesDS = require('../ds-responses');
+var responsesDS = require('../datasources/responses');
 
 var DEFAULT_BUFFER = 0.05;
 
@@ -231,7 +231,7 @@ exports.render = function render(req, res, next) {
     boundsBuffer: smartBuffer,
     surveyId: surveyId
   };
-  
+
   // It's useful to allow date filters as URL query paramters rather than
   // just as part of the layer definition. Most date filter scenarios will
   // involve "now", which is always changing. This avoids having each of those
@@ -296,7 +296,7 @@ exports.renderGrids = function renderGrids(req, res, next) {
     surveyId: surveyId,
     type: 'grid'
   };
-  
+
   if (until) {
     options.until = until;
   }
@@ -312,7 +312,7 @@ exports.renderGrids = function renderGrids(req, res, next) {
     }
     res.send(data);
   };
-  
+
   layer.getDefinition({
     layerId: layerId,
     // Used for legacy queries
