+++ lib/datasources/features.js
@@ -8,16 +8,13 @@ var knex = require('knex');
 var projector = require('nodetiles-core').projector;
 
 var metrics = require('../metrics/pipeline');
+var pg = require('../postgres').getClient();
+
 var settings = require('../settings');
 
 var maxResponseCount = settings.maxResponseCount;
 
 var FEATURE = 'Feature';
-
-var pg = require('knex')({
-  client: 'pg',
-  connection: settings.psqlConnectionString
-});
 
 function ToFeatures(options) {
   // Start and pause the timer here, in case we encounter 0 documents in the
