+++ lib/postgres.js
@@ -1,0 +1,12 @@
+var knex = require('knex');
+
+var settings = require('./settings');
+
+var pg = knex({
+  client: 'pg',
+  connection: settings.psqlConnectionString
+});
+
+exports.getClient = function getClient() {
+  return pg;
+};
