+++ lib/routes.js
@@ -5,15 +5,15 @@ var crypto = require('crypto');
 var bodyParser = require('body-parser');
 var knox = require('knox');
 
-var settings = require('./settings');
-var useEtagCache = require('./etag-cache')();
-var s3Cache = require('./s3-cache');
-var s3CacheFeatures = require('./s3-cache-features');
-var mongoCache = require('./mongo-cache');
-
 var features = require('./controllers/features');
+var mongoCache = require('./cache/mongo-cache');
+var pg = require('./postgres').getClient();
+var s3Cache = require('./cache/s3-cache');
+var s3CacheFeatures = require('./cache/s3-cache-features');
+var settings = require('./settings');
 var tiles = require('./controllers/tiles');
 var tileJson = require('./controllers/tile-json');
+var useEtagCache = require('./etag-cache')();
 
 var useCache, useCacheFeatures;
 
@@ -41,6 +41,7 @@ if (settings.cacheMethod === 'mongo') {
   });
 
   useCacheFeatures = s3CacheFeatures({
+    pgClient: pg,
     s3client: knoxClient
   });
 }
