+++ lib/settings.js
@@ -13,7 +13,7 @@ settings.s3Bucket = process.env.S3_BUCKET;
 
 // Postgresql parcel server
 // Use Heroku-style primary postgresql database environment variable
-settings.psqlConnectionString = process.env.POSTGRES_URL;
+settings.psqlConnectionString = process.env.DATABASE_URL;
 settings.featuresTable = process.env.FEATURES_TABLE;
 
 settings.port = process.env.PORT || process.argv[2] || 3001;
